Adds trajectory generator (WIP)
This was an attempt at adding TPS velocity & acceleration limits. The limiting logic kind of works, but at higher MAX_V (>160ish) it seems to glitch out and slap the throttle around. It's unfortunate that the math ended up using a lot of floating point variables, but it's a proof of concept ...